external modules: refactored some more sql utils modules, added fs (filesystem utils models)
