+++ external/native/tasks/soap_api.py
@@ -2,7 +2,7 @@
 from witness import Batch
 from witness.extractors.http import HttpGetExtractor
 from external.native.connections import Connection
-from external.utils.var import render_dump_path
+from external.utils.fs import render_dump_path
 
 
 def parse_soap_xml(rxml, config):
